+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Improved web UI indication for which data sources on fields have been overridden by data sources with higher precedence.
+- Added rake tasks for auditing configuration data sources.
+- Added rake tasks for documenting configuration fields and adding type definitions.
 
 ## 2.8.0